Make rec an optional arugment for undo_delete

+++ corehq/ex-submodules/dimagi/utils/couch/undo.py
@@ -41,7 +41,7 @@ class DeleteDocRecord(DeleteRecord):
 
     def undo(self):
         doc = self.get_doc()
-        undo_delete(self, doc)
+        undo_delete(doc, delete_record=self)
 
 
 class UndoableDocument(Document):
@@ -91,11 +91,12 @@ def get_deleted_doc_type(document_class_or_instance):
     return '{}{}'.format(base_name, DELETED_SUFFIX)
 
 
-def undo_delete(delete_record, document, save=True):
-    DeletedCouchDoc.objects.filter(
-        doc_id=delete_record._id,
-        doc_type=delete_record.doc_type,
-    ).delete()
+def undo_delete(document, delete_record=None, save=True):
+    if delete_record:
+        DeletedCouchDoc.objects.filter(
+            doc_id=delete_record._id,
+            doc_type=delete_record.doc_type,
+        ).delete()
     document.doc_type = remove_deleted_doc_type_suffix(document['doc_type'])
     if save:
         document.save()
